新增用户注册时邮件验证功能；
新增点赞量同步；
重构系统业务功能架构；
一系列系统优化与问题修复。

+++ src/main/java/com/wolfhouse/wolfhouseblog/service/impl/ArticleActionServiceImpl.java
@@ -24,13 +24,10 @@ import com.wolfhouse.wolfhouseblog.pojo.vo.ArticleBriefVo;
 import com.wolfhouse.wolfhouseblog.pojo.vo.ArticleCommentVo;
 import com.wolfhouse.wolfhouseblog.pojo.vo.ArticleFavoriteVo;
 import com.wolfhouse.wolfhouseblog.service.ArticleActionService;
-import com.wolfhouse.wolfhouseblog.service.ArticleService;
 import com.wolfhouse.wolfhouseblog.service.mediator.ServiceAuthMediator;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -51,17 +48,10 @@ import static com.wolfhouse.wolfhouseblog.pojo.domain.table.ArticleLikeTableDef.
 @Service
 @RequiredArgsConstructor
 public class ArticleActionServiceImpl implements ArticleActionService {
-    private ArticleService articleService;
     private final ServiceAuthMediator mediator;
     private final ArticleCommentMapper commentMapper;
     private final ArticleFavoriteMapper favoriteMapper;
     private final ArticleLikeMapper likeMapper;
-
-    @Autowired
-    @Qualifier("articleServiceImpl")
-    public void setArticleService(ArticleService articleService) {
-        this.articleService = articleService;
-    }
 
     @PostConstruct
     private void init() {
@@ -178,14 +168,24 @@ public class ArticleActionServiceImpl implements ArticleActionService {
                                              ArticleComment comment = commentMapper.selectOneById(commentId);
                                              return comment.getUserId()
                                                            .equals(t);
-                                         }))
+                                         })
+                                     .setCustomException(new ServiceException(ArticleConstant.NOT_ALLOWED)))
                   .doVerify();
 
         Set<Long> ids = getReplyIds(commentId);
-        int i = commentMapper.deleteBatchByIds(ids);
-        if (i != ids.size()) {
+        // 移除回复评论
+        if (!ids.isEmpty()) {
+            int i = commentMapper.deleteBatchByIds(ids);
+            if (i != ids.size()) {
+                throw new ServiceException(ArticleConstant.COMMENT_DELETE_FAILED);
+            }
+        }
+        // 移除本评论
+        int i = commentMapper.deleteById(commentId);
+        if (i != 1) {
             throw new ServiceException(ArticleConstant.COMMENT_DELETE_FAILED);
         }
+
         try {
             return getArticleCommentVosByArticle(dto.getArticleId());
         } catch (Exception e) {
@@ -228,7 +228,7 @@ public class ArticleActionServiceImpl implements ArticleActionService {
         if (isLiked(articleId)) {
             throw new ServiceException(ArticleConstant.ALREADY_LIKED);
         }
-        return likeMapper.insert(new ArticleLike(null, login, articleId, null)) == 1;
+        return likeMapper.insert(new ArticleLike(null, login, articleId, null), true) == 1;
     }
 
     @Override
@@ -284,7 +284,8 @@ public class ArticleActionServiceImpl implements ArticleActionService {
                                            .collect(Collectors.toSet());
 
         // 获取结果
-        List<ArticleBriefVo> brief = articleService.getBriefByIds(articleIds);
+        List<ArticleBriefVo> brief = mediator.articleService()
+                                             .getBriefByIds(articleIds);
         // 收藏记录分页结果的信息即为最终的分页结果信息
         Page<ArticleBriefVo> res = new Page<>(dto.getPageNumber(), dto.getPageSize(), favoritePage.getTotalRow());
         res.setRecords(brief);
@@ -356,5 +357,12 @@ public class ArticleActionServiceImpl implements ArticleActionService {
                         .and(ARTICLE_FAVORITE.USER_ID.eq(login)));
         log.info("清空收藏夹 {}，共 {} 条结果", favoritesId, i);
         return true;
+    }
+
+    @Override
+    public Long likeCount(Long id) {
+        return likeMapper.selectCountByQuery(
+            QueryWrapper.create()
+                        .where(ARTICLE_LIKE.ARTICLE_ID.eq(id)));
     }
 }
